Changes and New Template
Took the "bottom spacer" and converted it to a footer with a copyright symbol and placeholder text. Couldn't get the "git add" command to work so i'm doing this so i can go to bed. 
Also took the current index and duplicated it to make a template, i'd suggest working out the template in its entirity first, then you can simply paste text into the about, crew, contact, ect. Issue with it now is the white background box resizes with the text instead of staying one size, there's a comment in the HTML about it.